fix: Allow logs command to be used outside of Modmail threads
When the logs command is used with a UID argument, the current version
of Modmail requires the command to be run in a Modmail thread.
This behavior change appears to have been introduced with commit
9d4b8e3 as previous versions allowed the command to be used outside
of Modmail threads.

The other logs commands (closed-by, delete, key, responded, search) do
not require the command to be run in a thread.  The logs [user] command
should behave the same way.